Some minimal cleanup
There are a LOT of conversions from uint64_t to uint32_t that are NOT
sane and could overflow

+++ src/comp_types/base_packed_comp.h
@@ -62,8 +62,7 @@ template <class T>
 class BasePackedComponent {
 public:
 
-
-  void set_hacked(const uint64_t _old_size, const uint64_t _old_num_vars) {
+  void set_hacked(const uint32_t _old_size, const uint32_t _old_num_vars) {
     old_size = _old_size;
     old_num_vars = _old_num_vars;
     delete[] data_;
